# bubuka_pro/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OpenApi Documentation",
        "description": "Документация для микросервиса",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/platforms/{platform_id}/": {
            "get": {
                "tags": [
                    "Platforms"
                ],
                "summary": "Получение записи о платформе",
                "operationId": "1864e18d5cde7d9d5e2c1b82d28f1298",
                "parameters": [
                    {
                        "name": "platform_id",
                        "in": "path",
                        "description": "Id платформы",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 3
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Platform"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/404"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/401"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Platforms"
                ],
                "summary": "Удаление записи о платформе",
                "operationId": "ae45da8f2847c62d66db3765f27f9c83",
                "parameters": [
                    {
                        "name": "platform_id",
                        "in": "path",
                        "description": "Id платформы",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Статус ответа",
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/404"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/401"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Platforms"
                ],
                "summary": "Изменение записей о платформе ",
                "operationId": "c9be3603013177eade9f2090ee4312b6",
                "parameters": [
                    {
                        "name": "platform_id",
                        "in": "path",
                        "description": "Id проекта",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Название платформы",
                                        "type": "string",
                                        "example": "Windows 12"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Platform"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Статус кода",
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "errors": {
                                            "description": "Сообщение ошибки",
                                            "properties": {
                                                "name": {
                                                    "description": "Возможные ошибки валидации поля name",
                                                    "type": "string",
                                                    "example": [
                                                        "The name field is required.",
                                                        "The name field must be a string",
                                                        "The name field must not be greater than 255 characters.",
                                                        "The name has already been taken."
                                                    ]
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "code": {
                                            "description": "Код ответа",
                                            "type": "integer",
                                            "example": 422
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/404"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/401"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Platform": {
                "properties": {
                    "status": {
                        "type": "boolean",
                        "example": true
                    },
                    "body": {
                        "properties": {
                            "paltforms/platform": {
                                "properties": {
                                    "id": {
                                        "description": "Id платформы",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "name": {
                                        "description": "Название платформы",
                                        "type": "string",
                                        "example": "Ubuntu 20.04"
                                    },
                                    "deleted_at": {
                                        "description": "Дата удаления",
                                        "type": "string",
                                        "format": "date-time",
                                        "example": null,
                                        "nullable": true
                                    },
                                    "created_at": {
                                        "description": "Дата создания",
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2023-12-27T09:59:49.000000Z"
                                    },
                                    "updated_at": {
                                        "description": "Дата изменения",
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2023-12-27T09:59:49.000000Z"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "404": {
                "properties": {
                    "status": {
                        "description": "Статус ответа",
                        "type": "boolean",
                        "example": "false"
                    },
                    "message": {
                        "description": "Сообщение ответа",
                        "type": "string",
                        "example": "Запись не найдена"
                    },
                    "code": {
                        "description": "Код ответа",
                        "type": "integer",
                        "example": "404"
                    }
                },
                "type": "object"
            },
...
            "401": {
                "properties": {
                    "status": {
                        "description": "Статус ответа",
                        "type": "boolean",
                        "example": "false"
                    },
                    "message": {
                        "description": "Сообщение ответа",
                        "type": "string",
                        "example": [
                            "Token not provided",
                            "The token has been blacklisted"
                        ]
                    },
                    "code": {
                        "description": "Код ответа",
                        "type": "integer",
                        "example": "401"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "scheme": "bearer",
                "description": "Enter token",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}